column fixes
