+++ packages/sections/src/variant/VariantEffectPredictor/index.ts
@@ -3,5 +3,5 @@ export const definition = {
   id,
   name: "Variant Effect Predictor",
   shortName: "VE",
-  hasData: data => data,   // !! ONCE HAVE API, CHECK IF SIMPLY CHECKING FOR PRESENCE IS ENOUGH
+  hasData: data => data?.transcriptConsequences ?.length > 0,
 };
